Reverted The StaticFileServer to its original name
